Bugs, custom parameter

+++ src/application/gui/menus/params/baseparam.cc
@@ -24,7 +24,7 @@ void BaseParam::setDisplayPosition(uint8_t xCoord)
 	m_xDisplayPosition = xCoord;
 }
 
-void BaseParam::setScaling(uint8_t stepSize, uint8_t offset)
+void BaseParam::setScaling(uint8_t stepSize, int32_t offset)
 {
 	m_stepSize = stepSize;
 	m_offset = offset;
